requirements.txt
aiohttp
aioify
asyncio
async_timeout
bs4
discord.py
git+https://github.com/STFCcommunity/disputils.git
ety
gspread_asyncio
oauth2client
python-dotenv
wiktionaryparser
